backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
transformers==4.35.0
torch==2.1.0
nltk==3.8.1
numpy==1.24.3
scikit-learn==1.3.2
gunicorn==21.2.0
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
reportlab==4.0.7
Werkzeug==3.0.1
googletrans==4.0.0rc1
langdetect==1.0.9
sentencepiece==0.1.99
protobuf==3.20.3
PyJWT==2.8.0